fixed ui issues
