+++ src/components/Header/MainBlock/MainBlock.scss
@@ -9,7 +9,7 @@
   align-items: center;
   margin-bottom: 10px;
   flex-direction: column;
-  gap: 16px;
+  gap: 24px;
 
   @include onDesktop {
     justify-content: space-between;
