+++ src/components/Header/MainBlock/index.tsx
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom';
 import { ReactComponent as Logo } from '@assets/svg/logo.svg';
 
 import { Icons } from '../Icons';
-import { NavigationBar } from '../NavigationBar';
+import { SearchBar } from '../SearchBar';
 
 import './MainBlock.scss';
 
@@ -13,7 +13,7 @@ export const MainBlock = () => {
       <Link to="/" className="header-main__link">
         <Logo className="header-main__logo" />
       </Link>
-      <NavigationBar />
+      <SearchBar />
       <Icons />
     </div>
   );
