+++ src/components/Header/NavItem/NavItem.scss
@@ -3,19 +3,38 @@
 
 .nav-item {
   cursor: pointer;
+  display: flex;
+  align-items: center;
+  justify-content: center;
 
   &__wrapper {
     display: flex;
     align-items: center;
     position: relative;
+    justify-content: space-between;
+    width: 80%;
+    gap: 4px;
+
+    @include onDesktop {
+      width: 100%;
+    }
   }
 
   &__text {
     font-family: $font-poppins;
     color: $color-main-text;
     font-weight: 500;
+    font-size: 12px;
     line-height: 18px;
     padding: 18px 0;
+
+    @include onTablet {
+      font-size: 14px;
+    }
+
+    @include onDesktop {
+      font-size: 16px;
+    }
   }
 
   &__svg-box {
