+++ src/components/Header/Navigation/Navigation.scss
@@ -5,11 +5,22 @@
   width: 100%;
   background-color: $color-bg-placeholer;
   padding: 0 25px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  gap: 42px;
-  flex-wrap: wrap;
+
+  &__list {
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+
+    @include onTablet {
+      grid-template-columns: repeat(3, 1fr);
+    }
+
+    @include onDesktop {
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+      gap: 42px;
+    }
+  }
 
   @include onTablet {
     padding: 0 45px;
