--- src/components/Header/NavigationBar/NavigationBar.scss
@@ -1,65 +1,0 @@
-@import "../../../styles/_variables";
-@import "../../../styles/_mixins";
-
-.nav-bar {
-  height: 42px;
-  background-color: $color-bg-placeholer;
-  border: 1px solid $color-border-searchbar;
-  border-radius: 12px;
-  padding: 10px 5px 10px 8px;
-  font-family: inherit;
-  display: flex;
-  align-items: center;
-  gap: 4px;
-  max-width: 500px;
-
-  @include onTablet {
-    padding: 10px 18px 10px 16px;
-    width: 500px;
-    gap: 24px
-  }
-
-
-  &__dropdown {
-    @include flex-settings;
-    cursor: pointer;
-    height: 42px;
-    position: relative;
-
-    @include onTablet {
-      gap: 8px;
-    }
-  }
-
-  &__text {
-    font-weight: 700;
-    font-size: 12px;
-    line-height: 22px;
-    width: 100%;
-    text-align: center;
-
-    @include onDesktop {
-      font-size: 15px;
-    }
-  }
-
-  &__svg-box {
-    width: 16px;
-    height: 42px;
-  }
-
-  &__svg {
-    width: 100%;
-    height: 100%;
-    transition: transform 0.3s ease-in-out;
-
-    &--active {
-      transform: rotate(180deg);
-    }
-  }
-
-  &__line {
-    border: 1px solid $color-border-searchbar;
-    height: 22px;
-  }
-}
