+++ src/components/Header/SearchBar/SearchBar.scss
@@ -1,34 +1,65 @@
 @import "../../../styles/_variables";
 @import "../../../styles/_mixins";
 
-
 .search-bar {
-  width: 60%;
+  height: 42px;
+  background-color: $color-bg-placeholer;
+  border: 1px solid $color-border-searchbar;
+  border-radius: 12px;
+  padding: 10px 5px 10px 8px;
+  font-family: inherit;
   display: flex;
   align-items: center;
+  gap: 4px;
+  max-width: 500px;
 
-  &__input {
-    font-family: $font-sans;
-    flex-shrink: 1;
-    width: 100%;
-    outline: none;
-    border: none;
-    line-height: 18px;  
-    background-color: inherit;
+  @include onTablet {
+    padding: 10px 18px 10px 16px;
+    width: 500px;
+    gap: 24px
+  }
 
-    &::placeholder {
-      font-size: 12px;
-      color: $color-placeholder;
 
-      @include onTablet {
-        font-size: 14px;
-      }
+  &__dropdown {
+    @include flex-settings;
+    cursor: pointer;
+    height: 42px;
+    position: relative;
+
+    @include onTablet {
+      gap: 8px;
     }
   }
 
+  &__text {
+    font-weight: 700;
+    font-size: 12px;
+    line-height: 22px;
+    width: 100%;
+    text-align: center;
+
+    @include onDesktop {
+      font-size: 15px;
+    }
+  }
+
+  &__svg-box {
+    width: 16px;
+    height: 42px;
+  }
+
   &__svg {
-    padding-top: 2px;
-    @include square-var(16px);
-    cursor: pointer;
+    width: 100%;
+    height: 100%;
+    transition: transform 0.3s ease-in-out;
+
+    &--active {
+      transform: rotate(180deg);
+    }
+  }
+
+  &__line {
+    border: 1px solid $color-border-searchbar;
+    height: 22px;
   }
 }
