+++ src/components/Header/SearchInput/SearchInput.scss
@@ -1,0 +1,34 @@
+@import "../../../styles/_variables";
+@import "../../../styles/_mixins";
+
+
+.search {
+  width: 60%;
+  display: flex;
+  align-items: center;
+
+  &__input {
+    font-family: $font-sans;
+    flex-shrink: 1;
+    width: 100%;
+    outline: none;
+    border: none;
+    line-height: 18px;  
+    background-color: inherit;
+
+    &::placeholder {
+      font-size: 12px;
+      color: $color-placeholder;
+
+      @include onTablet {
+        font-size: 14px;
+      }
+    }
+  }
+
+  &__svg {
+    padding-top: 2px;
+    @include square-var(16px);
+    cursor: pointer;
+  }
+}
