+++ src/components/PageWrapper/PageWrapper.scss
@@ -7,4 +7,9 @@
   display: flex;
   flex-direction: column;
   min-height: 100vh;
+  padding: 0 10px;
+
+  @include onTablet {
+    padding: 0 45px;
+  }
 }
